Add ShaderProgram abstraction
ShaderProgram represents the programs that will be used by OpenGL when drawing. It is made from one or more ShaderObjects that will be linked.

Comes with Move Ctor/Ass and can be constructed from a vertex shader or from both a vertex and fragment shader.

Currently setting uniforms is not part of this class in case that should be managed elsewhere on a level not connected to each shader.